Reworked structure of views and enhanced admin, detail and deck change
