+++ cardDatabase/templates/tournament/tournament_admin.html
@@ -129,6 +129,11 @@
         <div class="row" id="playerList">
             
         </div>
+        <br>
+        <h4>Ruler breakdown</h4>
+        <br>
+        <div id='ruler-breakdown'>
+        <br>
         <div class="modal fade" id="playerRemoveModal" tabindex="-1" role="dialog" aria-labelledby="playerRemoveModal" aria-hidden="true">
         <div class="modal-dialog" role="document">
             <div class="modal-content">
@@ -144,7 +149,7 @@
                 </div>
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
-                    <button onclick="removePlayerFromTournament" data-dismiss="modal" type="button" class="btn btn-danger">Remove</button>
+                    <button id="remove-player-btn" data-dismiss="modal" type="button" class="btn btn-danger">Remove</button>
                 </div>
             </div>
         </div>
@@ -152,6 +157,11 @@
     </div>
 
     {% if staffAccount.role.can_delete %}
+        <h2>Staff Control</h2>
+        <div class="search-group">
+            <input id="staff-name-search" class="mr-3 form-control search-input" placeholder="Username"/>
+            <button id="staff-add" class="btn btn-primary">Add User</button>
+        </div>
         <div class="container mt-5">
             <button class="btn btn-danger" id="deleteBtn" data-toggle="modal" data-target="#deleteModal">Delete Tournament</button>
         </div>
@@ -185,6 +195,7 @@
     {{ block.super }}
     <script>
         window.canDelete = {% if staffAccount.role.can_delete %}true{% else %}false{% endif %};
+        rulers = 
     </script>
     <script type="text/javascript" src="{% static 'js/bootstrap/bootstrap.js' %}"></script>
     <script src="https://cdn.jsdelivr.net/npm/flatpickr"></script>
