+++ cardDatabase/templates/tournament/tournament_deck_change_denied.html
@@ -1,0 +1,19 @@
+{% extends 'cardDatabase/html/base.html' %}
+{% load card_database_tags static %}
+
+{% block css %}
+    {{ block.super }}
+    <link rel="stylesheet" href="{% static 'css/desktop_only.css' %}">
+{% endblock %}
+
+{% block body %}
+    <div>
+        <p>
+            {% block error_message %}Deck change is currently not possible!{% endblock %}
+        </p>
+        <p>
+            <button type="button"><a href="/">Home</a></button>
+        </p>
+        <img src="{% get_random_chibi 'sad' %}">
+    </div>
+{% endblock %}
