+++ cardDatabase/templates/tournament/tournament_detail.html
@@ -50,7 +50,25 @@
     </div>
     <hr>
     {% if currentPlayer is not None %}
+        <h4>Player Registration</h4>
         <strong>Current Registration Status:</strong> {{currentPlayer.registration_status}} </br>
+        {% if tournament.phase == 'registration' %}
+            <a href="{% url 'cardDatabase-tournament-change-decklist' tournament_id=tournament.pk %}" class="mt-3 mb-3 btn btn-primary">Change Decklist</a>
+            <a href="{% url 'cardDatabase-tournament-delete-registration' tournament_id=tournament.pk %}" class="mt-3 mb-3 btn btn-danger">Delete Registration</a>
+        {% endif %}
+        <h5 class="mt-3 mb-3">Current Deck:</h5>
+        <div class="col-md-4 mt-3 mb-3">
+            <div class="card">
+                <div class="card-body">
+                    <img class="card-img" src="{{ currentPlayer.deck.get_front_card_of_deck.card.card_image.url }}"></img>
+                    <h5 class="card-title">{{ currentPlayer.deck.name }}</h5>
+                    {% colours_to_imgs currentPlayer.deck.get_colours %}
+                    <div class="mb-2"><strong>Last Update:</strong> <span class="local-date-time" data-epoch="{% datetime_to_timestamp currentPlayer.deck.last_modified %}"></span></div>
+                    <a href="{% url 'cardDatabase-view-decklist' currentPlayer.deck.pk %}" target="_blank" class="btn btn-sm btn-info" target="_blank">View Decklist</a>
+                </div>
+            </div>
+        </div>
+        <hr>
     {% endif %}
     <h4 class="mb-3">Players</h4>
     {% if playerCount > 0 %}
@@ -63,6 +81,8 @@
                     <th>Dropped</th>
                     {% if tournament.reveal_decklists %}
                         <th>Decklist</th>
+                        <th>Rulers</th>
+                        <th>Will Attributes</th>
                     {% endif %}
                 </tr>
             </thead>
@@ -74,6 +94,18 @@
                         <td>{% if player.dropped_out %}Yes{% else %}No{% endif %}</td>
                         {% if tournament.reveal_decklists %}
                             <td><a href="{% url 'cardDatabase-view-decklist-share' decklist_id=player.deck.pk share_parameter=player.deck.shareCode %}" class="btn btn-sm btn-outline-primary">View</a></td>
+                            {% if player.deck.get_deck_rulers %}
+                                <td>
+                                    {% for ruler in player.deck.get_deck_rulers %}
+                                        <a href="{% url 'cardDatabase-view-card' ruler.card.card_id %}" target="_blank">
+                                            <img class="card-ruler-img" src="{{ ruler.card.card_image.url }}"/>
+                                        </a>
+                                    {% endfor %}
+                                </td>
+                            {% else %}
+                                <td></td>
+                            {% endif %}                            
+                            <td>{% colours_to_imgs player.deck.get_colours %}</td>
                         {% endif %}
                     </tr>
                 {% endfor %}
