+++ cardDatabase/views/tournament/admin/post/lock_deck_edit.py
@@ -1,10 +1,14 @@
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse, Http404
+from django.views.decorators.http import require_POST
 
-from ....models.Tournament import Tournament, TournamentStaff
+from .....models.Tournament import Tournament, TournamentStaff, TournamentPlayer
+
+from fowsim import constants as CONS
 
 @login_required
+@require_POST
 def post (request, tournament_id):
     if not request.method == "POST":
         raise Http404
@@ -18,10 +22,20 @@ def post (request, tournament_id):
     
     data = dict(request.POST)
     
-    reveal_state = 'revealState' in data
+    lock_state = 'lockState' in data
     
-    if tournament.reveal_decklists != reveal_state:
-        tournament.reveal_decklists = reveal_state
+    if tournament.deck_edit_locked != lock_state:
+        tournament.deck_edit_locked = lock_state
+        #update the lock state on each deck
         tournament.save()
+        
+        for player in TournamentPlayer.objects.filter(tournament=tournament):
+            if lock_state:
+                player.deck.deck_lock = CONS.MODE_TOURNAMENT
+            else:
+                player.deck.deck_lock = ''
+
+            player.deck.save()
+        
 
     return JsonResponse({'success': True})
