+++ cardDatabase/views/tournament/tournament_decklist.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def get(request):
+    return render(request, 'tournament/tournament_decklist_locked.html')
