+++ cardDatabase/views/tournament/tournament_remove_invalid.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def get(request):
+    return render(request, 'tournament/tournament_remove_invalid.html')
